Added uvm tb

+++ run.sh
@@ -1,17 +1,46 @@
 #!/bin/bash
-module load modelsim
-echo "	simulating... please wait"
-{
-  #run vsim
 
-  vsim -c -do ./scripts/3_simul.do
+if [ "$1" = "sv" ]
+then
 
-  #save log file
-  mv transcript simulation.log
+  # module load modelsim
+  echo "	simulating... please wait"
+  {
+    #run vsim
 
-  #cleanup
-  rm -rf *.ini transcript *.wlf work ./src/*~
+    vsim -c -do ./scripts/sv/3_simul.do
 
-} > /dev/null
-module unload modelsim
+    #save log file
+    mv transcript simulation.log
+
+    #cleanup
+    rm -rf *.ini transcript *.wlf work ./src/*~
+
+  } > /dev/null
+  # module unload modelsim
+
+elif [ "$1" = "uvm" ]
+then
+
+  # module load modelsim
+  echo "	simulating... please wait"
+  {
+
+    #run sim
+    vsim -c -do scripts/uvm/1_compile.do
+
+    #save log file
+    mv transcript simulation.log
+
+    #cleanup
+    rm -rf dump.vcd teste* transcript utopia.ucdb  vsim.wlf work *.ini
+
+  } > /dev/null
+  # module load modelsim
+
+else
+
+  echo "NOT AVAILABLE OPTION. Please, run \"./run.sh sv\" or \"./run.sh uvm\" "
+
+fi
